requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pydantic
pymysql
resend
numpy>=1.23.0
pandas>=1.5.0
tensorflow>=2.10.0
matplotlib>=3.5.0
pillow>=9.2.0
python-multipart
cryptography
httpx
aiofiles>=23.0.0